All around fixing and improving

+++ include/libplot_trace/PlotRelayPublisher.hpp
@@ -18,16 +18,23 @@ limitations under the License.
 #include <libcommon/libcommon.hpp>
 #include <libplot_trace/PlotRelayProtocol.hpp>
 
+/**
+ * @class PlotRelayPublisher
+ * @brief A final class responsible for publishing plot data points to the plot relay.
+ *
+ * This class provides multiple overloaded `publish` methods to send data
+ * for up to four curves to a plot relay system. It is non-copyable.
+ */
 class PlotRelayPublisher final : NonCopyable
 {
     public:
         PlotRelayPublisher();
         ~PlotRelayPublisher();
 
-        void publish(const char* graph_name, uint32_t graph_name_len, float stamp, const char* curve_1, uint32_t curve_1_len, float value_1);
-        void publish(const char* graph_name, uint32_t graph_name_len, float stamp, const char* curve_1, uint32_t curve_1_len, float value_1, const char* curve_2, uint32_t curve_2_len, float value_2);
-        void publish(const char* graph_name, uint32_t graph_name_len, float stamp, const char* curve_1, uint32_t curve_1_len, float value_1, const char* curve_2, uint32_t curve_2_len, float value_2, const char* curve_3, uint32_t curve_3_len, float value_3);
-        void publish(const char* graph_name, uint32_t graph_name_len, float stamp, const char* curve_1, uint32_t curve_1_len, float value_1, const char* curve_2, uint32_t curve_2_len, float value_2, const char* curve_3, uint32_t curve_3_len, float value_3, const char* curve_4, uint32_t curve_4_len, float value_4);
+        void publish(const char* graph_name, size_t graph_name_len, float stamp, const char* curve_1, size_t curve_1_len, float value_1);
+        void publish(const char* graph_name, size_t graph_name_len, float stamp, const char* curve_1, size_t curve_1_len, float value_1, const char* curve_2, size_t curve_2_len, float value_2);
+        void publish(const char* graph_name, size_t graph_name_len, float stamp, const char* curve_1, size_t curve_1_len, float value_1, const char* curve_2, size_t curve_2_len, float value_2, const char* curve_3, size_t curve_3_len, float value_3);
+        void publish(const char* graph_name, size_t graph_name_len, float stamp, const char* curve_1, size_t curve_1_len, float value_1, const char* curve_2, size_t curve_2_len, float value_2, const char* curve_3, size_t curve_3_len, float value_3, const char* curve_4, size_t curve_4_len, float value_4);
 
     private:
         int     _fd;
